+++ MeetBooking/celery.py
@@ -2,7 +2,7 @@ import os
 from celery import Celery
 from celery.schedules import crontab
 
-from MeetBooking.settings import STEP_TIME_MINUTES
+from django.conf import settings
 
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'MeetBooking.settings')
 
@@ -14,6 +14,6 @@ app.autodiscover_tasks()
 app.conf.beat_schedule = {
     'send-email': {
         'task': 'booking.tasks.send_events_mail',
-        'schedule': crontab(minute=f'*/{STEP_TIME_MINUTES}')
+        'schedule': crontab(minute=f'*/{settings.STEP_TIME_MINUTES}')
     }
 }
