--- booking/api_views/__init__.py
@@ -1,2 +1,0 @@
-from .cabinet_view_set import CabinetViewSet
-from .event_view_set import EventViewSet
