+++ booking/forms.py
@@ -46,7 +46,7 @@ class EventForm(forms.ModelForm):
 
     class Meta:
         model = Event
-        fields = ('title', 'start_time', 'end_time')
+        fields = ('title', 'start_time', 'end_time', 'visitors')
 
     def clean(self):
         data = self.cleaned_data
