+++ booking/templates/cabinets/carousel_item.html
@@ -1,0 +1,14 @@
+<div class="col">
+    <ul class="datetime-ul">
+        {% for date,times in date_time.items %}
+            <li class="date-li">{{ date }}</li>
+            {% for time,event in times.items %}
+                {% if event %}
+                    <li class="time-li open">{{ time }}</li>
+                {% else %}
+                    <li class="time-li close">{{ time }}</li>
+                {% endif %}
+            {% endfor %}
+        {% endfor %}
+    </ul>
+</div>
