+++ booking/urls.py
@@ -1,13 +1,14 @@
 from django.urls import path
 from rest_framework.routers import DefaultRouter
 
-from booking import api_views, views
+from booking.api.views import CabinetViewSet, EventViewSet
+from booking.views import CabinetDetailView
 
 router = DefaultRouter()
-router.register("api/cabinets", api_views.CabinetViewSet, basename="cabinets")
-router.register("api/cabinets/(?P<room_number>[^/.]+)/events", api_views.EventViewSet, basename="events")
+router.register("api/cabinets", CabinetViewSet, basename="cabinets")
+router.register("api/cabinets/(?P<room_number>[^/.]+)/events", EventViewSet, basename="events")
 urlpatterns = router.urls
 
 urlpatterns += [
-    path("cabinets/<int:pk>/", views.CabinetDetailView.as_view(), name="booking")
+    path("cabinets/<int:pk>/", CabinetDetailView.as_view(), name="booking")
 ]
