--- booking/views/__init__.py
@@ -1,1 +1,0 @@
-from .cabinet_detail_view import CabinetDetailView
