+++ requirements.txt
@@ -1,19 +1,9 @@
-amqp==2.6.1
-asgiref==3.4.1
-billiard==3.6.4.0
 celery==4.4.6
 Django==3.2.7
 django-debug-toolbar==3.2.2
 djangorestframework==3.12.4
 flake8==3.9.2
-future==0.18.2
-kombu==4.6.11
-mccabe==0.6.1
 psycopg2-binary==2.9.1
-pycodestyle==2.7.0
-pyflakes==2.3.1
 python-dotenv==0.19.0
-pytz==2021.1
 redis==3.5.3
-sqlparse==0.4.2
-vine==1.3.0
+
