+++ templates/base.html
@@ -2,6 +2,7 @@
 <html lang="ru">
 {% load static %}
 <head>
+    <meta charset="utf-8">
     <title> {% block title %} {% endblock %}</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/css/bootstrap.min.css" rel="stylesheet"
           integrity="sha384-+0n0xVW2eSR5OomGNYDnhzAbDsOXxcvSN1TPprVMTNDbiYZCxYbOOl7+AMvyTG2x" crossorigin="anonymous">
